separate regions DB object from type definition, add deferred loader
