+++ src/common/regions/index.ts
@@ -1,4 +1,4 @@
-import regions from 'common/regions/region_db';
+//import regions from 'common/regions/region_db';
 import {
   getStateName,
   getStateCode,
@@ -6,8 +6,8 @@ import {
   getFormattedStateCode,
 } from './regions_data';
 
-export default regions;
 export { getStateName, getStateCode, getStateFips, getFormattedStateCode };
+export * from './getRegionDB';
 export * from './types';
 export * from './region_hooks';
 export * from './utils';
