+++ src/components/EmailAlertsFooter/EmailAlertsFooter.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import EmailAlertsFooter from './EmailAlertsFooter';
-import regions from 'common/regions';
+import regions from 'common/regions/region_db';
 
 export default {
   title: 'Shared Components/EmailAlertsFooter',
