+++ src/components/Map/USACountyMap.tsx
@@ -6,7 +6,8 @@ import ReactTooltip from 'react-tooltip';
 import { colorFromLocationSummary } from 'common/colors';
 import { useSummaries } from 'common/location_summaries';
 import { ScreenshotReady } from 'components/Screenshot';
-import regions, { State as StateType } from 'common/regions';
+import regions from 'common/regions/region_db';
+import { State as StateType } from 'common/regions';
 import stateGeographies from 'common/data/states-10m.json';
 import { USMapWrapper, USStateMapWrapper } from './Map.style';
 import { trackEvent, EventAction, EventCategory } from 'components/Analytics';
