+++ src/components/RegionMap/RegionMap.tsx
@@ -4,7 +4,8 @@ import { ComposableMap, Geographies } from 'react-simple-maps';
 import ReactTooltip from 'react-tooltip';
 import * as topojson from 'topojson-client';
 import { geoBounds, geoCentroid, geoDistance } from 'd3-geo';
-import regions, { County, MetroArea, State, Region } from 'common/regions';
+import regions from 'common/regions/region_db';
+import { County, MetroArea, State, Region } from 'common/regions';
 import { LocationSummariesByFIPS } from 'common/location_summaries';
 import { useCountyGeographies, useStateGeographies } from 'common/hooks';
 import * as Styles from './RegionMap.style';
