+++ src/components/VulnerabilitiesBlock/FooterLinks/FooterLinks.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import FooterLinks from './FooterLinks';
-import regions from 'common/regions';
+import regions from 'common/regions/region_db';
 
 export default {
   title: 'Shared Components/VulnerabilitiesFooterLinks',
