+++ src/screens/HomePage/utils.ts
@@ -1,7 +1,7 @@
 import sortBy from 'lodash/sortBy';
 import reverse from 'lodash/reverse';
 import take from 'lodash/take';
-import regions from 'common/regions';
+import regions from 'common/regions/region_db';
 
 /* Gets fips of 5 largest metro areas to set as homepage explore presets */
 export function getLargestMetroFipsForExplore(numRegions: number) {
